fix: useTheme hook created, commonStyles variable, styles changed. Fixed a bug with buttons on the main screen.
