+++ packages/client/src/hooks/index.ts
@@ -5,6 +5,7 @@ export * from './useLogin'
 export * from './useProfile'
 export * from './useRegistration'
 export * from './useRoutes'
+export * from './useTheme'
 export * from './useUrl'
 export * from './useLogout'
 export * from './useFullScreen'
